Address review feedback for vertex enumeration split
