+++ crates/library/src/geom/vertex_enumeration/linear_algebra.rs
@@ -1,0 +1,121 @@
+//! Exact low-level linear algebra helpers over Q.
+
+use num_rational::BigRational;
+use num_traits::Zero;
+
+#[cfg(test)]
+fn det3(r0: &[BigRational], r1: &[BigRational], r2: &[BigRational]) -> BigRational {
+    &r0[0] * (&r1[1] * &r2[2] - &r1[2] * &r2[1]) - &r0[1] * (&r1[0] * &r2[2] - &r1[2] * &r2[0])
+        + &r0[2] * (&r1[0] * &r2[1] - &r1[1] * &r2[0])
+}
+
+#[cfg(test)]
+pub(super) fn det4(rows: &[[BigRational; 4]; 4]) -> BigRational {
+    let (a, b, c, d) = (&rows[0], &rows[1], &rows[2], &rows[3]);
+    let c00 = det3(
+        &[b[1].clone(), b[2].clone(), b[3].clone()],
+        &[c[1].clone(), c[2].clone(), c[3].clone()],
+        &[d[1].clone(), d[2].clone(), d[3].clone()],
+    );
+    let c01 = det3(
+        &[b[0].clone(), b[2].clone(), b[3].clone()],
+        &[c[0].clone(), c[2].clone(), c[3].clone()],
+        &[d[0].clone(), d[2].clone(), d[3].clone()],
+    );
+    let c02 = det3(
+        &[b[0].clone(), b[1].clone(), b[3].clone()],
+        &[c[0].clone(), c[1].clone(), c[3].clone()],
+        &[d[0].clone(), d[1].clone(), d[3].clone()],
+    );
+    let c03 = det3(
+        &[b[0].clone(), b[1].clone(), b[2].clone()],
+        &[c[0].clone(), c[1].clone(), c[2].clone()],
+        &[d[0].clone(), d[1].clone(), d[2].clone()],
+    );
+    &a[0] * c00 - &a[1] * c01 + &a[2] * c02 - &a[3] * c03
+}
+
+#[cfg(test)]
+pub(super) fn solve4(
+    rows: &[[BigRational; 4]; 4],
+    rhs: &[BigRational; 4],
+) -> Option<[BigRational; 4]> {
+    let d = det4(rows);
+    if d.is_zero() {
+        return None;
+    }
+    let mut result: [BigRational; 4] = std::array::from_fn(|_| BigRational::zero());
+    for col in 0..4 {
+        let mut modified = rows.clone();
+        for row in 0..4 {
+            modified[row][col] = rhs[row].clone();
+        }
+        result[col] = det4(&modified) / &d;
+    }
+    Some(result)
+}
+
+#[cfg(test)]
+pub(super) fn dot4(a: &[BigRational; 4], b: &[BigRational; 4]) -> BigRational {
+    &a[0] * &b[0] + &a[1] * &b[1] + &a[2] * &b[2] + &a[3] * &b[3]
+}
+
+#[cfg(test)]
+pub(super) fn cross_product_4d_rational(
+    a: &[BigRational; 4],
+    b: &[BigRational; 4],
+    c: &[BigRational; 4],
+) -> [BigRational; 4] {
+    let bc_01 = &b[0] * &c[1] - &b[1] * &c[0];
+    let bc_02 = &b[0] * &c[2] - &b[2] * &c[0];
+    let bc_03 = &b[0] * &c[3] - &b[3] * &c[0];
+    let bc_12 = &b[1] * &c[2] - &b[2] * &c[1];
+    let bc_13 = &b[1] * &c[3] - &b[3] * &c[1];
+    let bc_23 = &b[2] * &c[3] - &b[3] * &c[2];
+
+    let d0 = &a[1] * &bc_23 - &a[2] * &bc_13 + &a[3] * &bc_12;
+    let d1 = -(&a[0] * &bc_23 - &a[2] * &bc_03 + &a[3] * &bc_02);
+    let d2 = &a[0] * &bc_13 - &a[1] * &bc_03 + &a[3] * &bc_01;
+    let d3 = -(&a[0] * &bc_12 - &a[1] * &bc_02 + &a[2] * &bc_01);
+    [d0, d1, d2, d3]
+}
+
+pub(super) fn rank_over_q(rows: &[[BigRational; 4]]) -> usize {
+    if rows.is_empty() {
+        return 0;
+    }
+    let m = rows.len();
+    let mut mat: Vec<[BigRational; 4]> = rows.to_vec();
+    let mut rank = 0;
+    for col in 0..4 {
+        let Some(pivot_row) = (rank..m).find(|&r| !mat[r][col].is_zero()) else {
+            continue;
+        };
+        mat.swap(rank, pivot_row);
+        let pivot_val = mat[rank][col].clone();
+        for r in 0..m {
+            if r == rank || mat[r][col].is_zero() {
+                continue;
+            }
+            let factor = &mat[r][col] / &pivot_val;
+            let pivot = mat[rank].clone();
+            for (mat_c, pivot_c) in mat[r][col..4].iter_mut().zip(pivot[col..4].iter()) {
+                *mat_c = &*mat_c - &factor * pivot_c;
+            }
+        }
+        rank += 1;
+    }
+    rank
+}
+
+pub(super) fn affine_rank_rational(points: &[[BigRational; 4]]) -> usize {
+    if points.len() <= 1 {
+        return 0;
+    }
+    let base = &points[0];
+    let centered: Vec<[BigRational; 4]> = points[1..]
+        .iter()
+        .map(|p| std::array::from_fn(|i| &p[i] - &base[i]))
+        .collect();
+    rank_over_q(&centered)
+}
