+++ crates/library/src/geom/vertex_enumeration/tests/mod.rs
@@ -1,0 +1,4 @@
+//! Unit tests for the vertex enumeration pipeline.
+
+mod linear_algebra;
+mod pipeline;
